backend/requirements.txt
# LiveKit Core & Agents (MATCHED)
livekit==1.0.23
livekit-agents==1.3.11
livekit-api==1.1.0

# Plugins
livekit-plugins-google==1.3.11
livekit-plugins-silero==1.3.11
livekit-plugins-deepgram      # Free STT
livekit-plugins-cartesia       # Free TTS

llama-index>=0.12.0
llama-index-embeddings-gemini>=0.3.0
llama-index-llms-gemini>=0.4.0

docx2txt==0.8
PyPDF2==3.0.1
google-genai==0.19.0

# Web Server
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0

